+++ R/message.R
@@ -2,48 +2,20 @@
 
 #' Conditionally format messages
 #' 
-#' @param x message template to be passed to [stringi::stri_sprintf()].
-#' @param values list of values to be used in the template. Coerced to list if vector is given.
-#' @param indices list of integer to specify which value to be used.
+#' @inheritParams stringi::stri_sprintf()
 #' @param pretty if `TRUE`, message is passed to [base::prettyNum()].
-#' @param ... additional arguments passed to [base::prettyNum()].
 #' @keywords internal development
 #' @examples 
-#' \dontrun{
-#' quanteda:::msg("you cannot delete %s", 
-#'                c("a document", "documents"), indices = TRUE)
-#' quanteda:::msg("tokens has %s", 
-#'                c("sentences", "paragraphs", "documents"), indices = 2)
-#' 
-#' dfmat <- data_dfm_lbgexample
-#' quanteda:::msg("dfm has %d %s and %d %s", 
-#'      list(ndoc(dfmat), c("document", "documents"),
-#'           nfeat(dfmat), c("feature", "features")), 
-#'      list(1, ndoc(dfmat) > 1, 
-#'           1, nfeat(dfmat) > 1))
-#' }      
-msg <- function(x, values = NULL, indices = NULL, pretty = TRUE, ...) {
-    if (!is.null(values)) {
-        if (!is.list(values))
-            values <- list(values)
-        if (!is.list(indices))
-            indices <- as.list(indices)
-        if (length(indices)) {
-            values <- mapply(function(x, y) {
-                if (length(x) == 1 || is.null(y))
-                    return(x)
-                if (is.logical(y))
-                    y <- which(c(FALSE, TRUE) == y)
-                return(x[y])
-            }, values, indices, SIMPLIFY = FALSE)
-        }
-        msg <- do.call(stringi::stri_sprintf, c(list(x), values))
+#' quanteda:::msg("you cannot delete %s %s", 2000, "documents")
+msg <- function(format, ..., pretty = TRUE) {
+    args <- list(...)
+    if (pretty) {
+        args <- lapply(args, prettyNum, big.mark = ",")
     } else {
-        msg <- x
+        args <- lapply(args, as.character)
     }
-    if (pretty)
-        msg <- prettyNum(msg, big.mark = ",", ...)
-    return(msg)
+    args$format <- format
+    do.call(stringi::stri_sprintf, args)
 }
 
 # rdname catm
@@ -53,34 +25,30 @@ catm <- function(..., sep = " ", appendLF = FALSE) {
     message(paste(..., sep = sep), appendLF = appendLF)
 }
 
-# used in displaying verbose messages for tokens_select and dfm_select
-message_select <- function(selection, nfeats, ndocs, nfeatspad = 0, ndocspad = 0) {
-    catm(if (selection == "keep") "kept" else "removed", " ",
-         format(nfeats, big.mark = ",", scientific = FALSE),
-         " feature", if (nfeats != 1L) "s" else "", sep = "")
-    if (ndocs > 0) {
-        catm(" and ",
-             format(ndocs, big.mark = ",", scientific = FALSE),
-             " document", if (ndocs != 1L) "s" else "",
-             sep = "")
-    }
-    if ((nfeatspad + ndocspad) > 0) {
-        catm(", padded ", sep = "")
-    }
-    if (nfeatspad > 0) {
-        catm(format(nfeatspad, big.mark = ",", scientific = FALSE),
-             " feature", if (nfeatspad != 1L) "s" else "",
-             sep = "")
-    }
-    if (ndocspad > 0) {
-        if (nfeatspad > 0) catm(" and ", sep = "")
-        catm(format(ndocspad, big.mark = ",", scientific = FALSE),
-             " document", if (ndocspad != 1L) "s" else "",
-             sep = "")
-    }
-    catm("", appendLF = TRUE)
+# used in displaying verbose messages for tokens and dfm constructors
+message_create <- function(input, output) {
+    message(msg("Creating a %s from a %s object...",
+                output, input))
 }
 
+message_finish <- function(x, time) {
+    if (is.dfm(x)) {
+        message(msg(" ...complete, elapsed time: %s seconds.",
+                    format((proc.time() - time)[3], digits = 3)))
+        message(msg("Finished constructing a %s x %s sparse dfm.",
+                    nrow(x), ncol(x)))
+    } else {
+        m <- length(types(x))
+        n <- ndoc(x)
+        message(msg(" ...%s unique %s", 
+                    m, if (m == 1) "type" else "types"))
+        message(msg(" ...complete, elapsed time: %s seconds.",
+                    format((proc.time() - time)[3], digits = 3)))
+        message(msg("Finished constructing %s from %s %s",
+                    class(x)[1], 
+                    n, if (n == 1) "document" else "documents"))
+    }
+}
 
 # messaging methods ------------
              
@@ -99,10 +67,8 @@ NULL
 #' @inheritParams messages
 #' @keywords message internal
 message_tokens <- function(operation, before, after) {
-    msg <- sprintf("%s changed from %d tokens (%d documents) to %d tokens (%d documents)",
-                   operation, before$ntoken, before$ndoc, after$ntoken, after$ndoc)
-    msg <- prettyNum(msg, big.mark = ",")
-    message(msg)
+    message(msg("%s changed from %s tokens (%s documents) to %s tokens (%s documents)",
+                operation, before$ntoken, before$ndoc, after$ntoken, after$ndoc))
 }
 
 stats_tokens <- function(x) {
@@ -114,10 +80,8 @@ stats_tokens <- function(x) {
 #' @inheritParams messages
 #' @keywords message internal
 message_dfm <- function(operation, before, after) {
-    msg <- sprintf("%s changed from %d features (%d documents) to %d features (%d documents)",
-                   operation, before$nfeat, before$ndoc, after$nfeat, after$ndoc)
-    msg <- prettyNum(msg, big.mark = ",")
-    message(msg)
+    message(msg("%s changed from %s features (%s documents) to %s features (%s documents)",
+                operation, before$nfeat, before$ndoc, after$nfeat, after$ndoc))
 }
 
 stats_dfm <- function(x) {