+++ R/tokens-methods.R
@@ -69,13 +69,13 @@ print.tokens <- function(x, max_ndoc = quanteda_options("print_tokens_max_ndoc")
         max_ndoc <- ndoc(x)
 
     if (show_summary) {
-        cat("Tokens consisting of ", format(ndoc, big.mark = ","), " document",
-            if (ndoc != 1L) "s" else "", sep = "")
+        cat(msg("Tokens consisting of %s %s",
+                ndoc, if (ndoc == 1) "document" else "documents"))
         if (ncol(docvars))
-            cat(" and ", format(ncol(docvars), big.mark = ","), " docvar",
-                if (ncol(docvars) != 1L) "s" else "", sep = "")
+            cat(msg(" and %s %s",
+                    ncol(docvars), if (ncol(docvars) == 1) "docvar" else "docvars"))
         if (is.tokens_xptr(x))
-            cat(" (pointer to ", address(x), ")", sep = "")
+            cat(msg(" (pointer to %s)", address(x)))
         cat(".\n")
     }
 
@@ -91,13 +91,14 @@ print.tokens <- function(x, max_ndoc = quanteda_options("print_tokens_max_ndoc")
             cat(label[i], "\n", sep = "")
             print(x[[i]], ...)
             if (len[i] > max_ntoken)
-                cat("[ ... and ",  format(len[i] - max_ntoken, big.mark = ","), " more ]\n", sep = "")
+                cat(msg("[ ... and %s more ]\n", 
+                        len[i] - max_ntoken))
             cat("\n", sep = "")
         }
         ndoc_rem <- ndoc - max_ndoc
         if (ndoc_rem > 0)
-            cat("[ reached max_ndoc ... ", format(ndoc_rem, big.mark = ","), " more document",
-                if (ndoc_rem > 1) "s", " ]\n", sep = "")
+            cat(msg("[ reached max_ndoc ... %s more %s ]",
+                    ndoc_rem, if (ndoc_rem == 1) "document" else "documents"))
     }
 }